requirements.txt
requests_toolbelt==0.9.1
requests==2.20.0
allure_python_commons==2.8.5
pytest==5.2.1
simplejson==3.16.0
config==0.4.2
PyYAML==5.1.2
